Fixed the metric calculations

+++ cmd/migrate/migrations/20240423202949_add_table_metrics.up.sql
@@ -1,9 +1,9 @@
 CREATE TABLE IF NOT EXISTS metrics (
    id INTEGER PRIMARY KEY,
    user_id TEXT,
-   mrr INTEGER,
-   churn INTEGER,
+   mrr REAL,
+   churn REAL,
    net_growth INTEGER,
-   trading_limit INTEGER,
+   trading_limit REAL,
    last_updated TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
