+++ types/metric.go
@@ -8,9 +8,9 @@ import (
 type Metric struct {
 	ID           int       `bun:"id,pk,autoincrement"`
 	User         uuid.UUID `bun:"user_id,notnull"`
-	MRR          int64     `bun:"mrr,notnull"`
-	Churn        int       `bun:"churn,notnull"`
-	NetGrowth    int       `bun:"net_growth,notnull"`
-	TradingLimit int64     `bun:"trading_limit,notnull"`
+	MRR          float64   `bun:"mrr,notnull"`
+	Churn        float64   `bun:"churn,notnull"`
+	NetGrowth    int64     `bun:"net_growth,notnull"`
+	TradingLimit float64   `bun:"trading_limit,notnull"`
 	LastUpdated  time.Time `bun:"last_updated"`
 }
